+++ sites/partners/lib/hooks.ts
@@ -9,6 +9,7 @@ import {
   EnumListingFilterParamsComparison,
   EnumPreferencesFilterParamsComparison,
   EnumProgramsFilterParamsComparison,
+  EnumUserFilterParamsComparison,
 } from "@bloom-housing/backend-core/types"
 
 interface PaginationProps {
@@ -353,13 +354,12 @@ export function useUserList({ page, limit }: UseUserListProps) {
     userService.list({
       page,
       limit,
-      // TODO: temporary disabled, because it occurs an issue with data fetching (501 - Filter Not Implemented)
-      // filter: [
-      //   {
-      //     isPartner: true,
-      //     $comparison: EnumUserFilterParamsComparison["="],
-      //   },
-      // ],
+      filter: [
+        {
+          isPartner: true,
+          $comparison: EnumUserFilterParamsComparison["="],
+        },
+      ],
     })
 
   const { data, error } = useSWR(