+++ ui-components/src/page_components/listing/listing_sidebar/SubmitApplication.tsx
@@ -22,14 +22,18 @@ export interface ApplicationAddressesProps {
 const SubmitApplication = (props: ApplicationAddressesProps) => {
   return (
     <>
-      {(props.applicationMailingAddress || props.applicationDropOffAddress) && (
+      {(props.applicationMailingAddress ||
+        props.applicationDropOffAddress ||
+        props.postmarkedApplicationData) && (
         <section className="aside-block is-tinted bg-gray-100">
           <NumberedHeader num={2} text={t("listings.apply.submitAPaperApplication")} />
-          {props.applicationMailingAddress && (
+          {(props.applicationMailingAddress || props.postmarkedApplicationData) && (
             <>
               <h3 className="text-caps-tiny">{t("listings.apply.sendByUsMail")}</h3>
               <p className="text-gray-700">{props.applicationOrganization}</p>
-              <SidebarAddress address={props.applicationMailingAddress} />
+              {props.applicationMailingAddress && (
+                <SidebarAddress address={props.applicationMailingAddress} />
+              )}
               <p className="mt-4 text-tiny text-gray-750">
                 {props.postmarkedApplicationData?.postmarkedApplicationsReceivedByDate
                   ? t("listings.apply.postmarkedApplicationsMustBeReceivedByDate", {